Added a Pause Menu

+++ Endeavor/Assets/Scripts/PlayerData.cs
@@ -21,8 +21,8 @@ public class PlayerData
         inventory = player.inventory;
 
         position = new float[3];
-        position[0] = player.transform.position.x;
-        position[1] = player.transform.position.y;
-        position[2] = player.transform.position.z;
+        position[0] = (int)player.transform.position.x;
+        position[1] = (int)player.transform.position.y;
+        position[2] = (int)player.transform.position.z;
     }
 }
